packages/suite/src/utils/wallet/accountUtils.ts
import BigNumber from 'bignumber.js';
import { NETWORKS } from '@suite-config';
import { Account } from '@wallet-types';

export const parseBIP44Path = (path: string) => {
    const regEx = /m\/(\d+'?)\/(\d+'?)\/(\d+'?)\/([0,1])\/(\d+)/;
    const tokens = path.match(regEx);
    if (!tokens || tokens.length !== 6) {
        return null;
    }
    return {
        purpose: tokens[1],
        coinType: tokens[2],
        account: tokens[3],
        change: tokens[4],
        addrIndex: tokens[5],
    };
};

export const formatAmount = (amount: string, symbol: string) => {
    const network = NETWORKS.find(n => n.symbol === symbol);
    if (!network) return amount;
    return new BigNumber(amount).div(10 ** network.decimals).toString(10);
};

export const sortByCoin = (accounts: Account[]) => {
    return accounts.sort((a, b) => {
        const aIndex = NETWORKS.findIndex(n => {
            const accountType = n.accountType || 'normal';
            return accountType === a.accountType && n.symbol === a.symbol;
        });
        const bIndex = NETWORKS.findIndex(n => {
            const accountType = n.accountType || 'normal';
            return accountType === b.accountType && n.symbol === b.symbol;
        });
        if (aIndex === bIndex) return a.index - b.index;
        return aIndex - bIndex;
    });
};
